kernel: 32bit: fix capability tracing

+++ kernel/capabilities.c
@@ -34,11 +34,11 @@
 #include <bitmacros.h>
 
 #ifdef TRACE_PMEM_CAPS
-uintptr_t   trace_types_enabled = TRACE_TYPES_ENABLED_INITIAL;
-genpaddr_t TRACE_PMEM_BEGIN     = TRACE_PMEM_BEGIN_INITIAL;
-gensize_t  TRACE_PMEM_SIZE      = TRACE_PMEM_SIZE_INITIAL;
+uint64_t   trace_types_enabled = TRACE_TYPES_ENABLED_INITIAL;
+genpaddr_t TRACE_PMEM_BEGIN    = TRACE_PMEM_BEGIN_INITIAL;
+gensize_t  TRACE_PMEM_SIZE     = TRACE_PMEM_SIZE_INITIAL;
 
-void caps_trace_ctrl(uintptr_t types, genpaddr_t start, gensize_t size)
+void caps_trace_ctrl(uint64_t types, genpaddr_t start, gensize_t size)
 {
     if (types) {
         trace_types_enabled = types;
